Fix auto-routing to run under aspire deploy (annotation-driven)
BeforePublishEvent is only dispatched by PipelineExecutor (the legacy
`dotnet run -- --publisher kubernetes` path); `aspire deploy` invokes
pipeline steps directly over the CLI JSON-RPC backchannel and never
raises that event, which meant the WithClusterDefaults auto-router was
silently no-opping during real deploys and the Gateway/HTTPRoute
templates never got emitted.

Move auto-routing into ProcessGatewayResources itself, driven by a new
KubernetesGatewayAutoRouteAnnotation that WithClusterDefaults attaches
to the gateway. ProcessGatewayResources runs as part of the
prepare-deployment-targets pipeline step, which fires under every
execution path, so routes are now populated deterministically.
